+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.9...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.10...HEAD) - TBD
+
+## [4.21.10](https://github.com/schemathesis/schemathesis/compare/v4.21.9...v4.21.10) - 2026-06-20
 
 ### :bug: Fixed